# assets/mes/services.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "services.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/mes/api/v2/sign": {
      "post": {
        "summary": "作業員登入出機台工位",
        "description": "指定 in 的值表示為登入或登出. 登入: in=true; 登出: in=false.\n欲登入出的作業員身分指定於 metadata 的 User-Id, 而登入出的機台工位則分別依據 metadata 的\n Station 和 Site.\n\n登入時必需指定:\n - workDate: 工作日, unix nano格式.\n - group: 工作群組, 將影響生產批號(詳細請參考 rpc Collect).\n\n登出時僅需指定: in=false.\n\n如果 onlyValidation 指定為 true, 表示檢查該工號是否為在職.\n\n下列為回傳錯誤清單:\n - Error_ERROR_BAD_REQUEST\n - Error_ERROR_USER_NOT_FOUND_OR_BAD_PASSWORD (使用者已離職或密碼錯誤)\n - Error_ERROR_US_MISMATCH (作業員與機台登入人員不符, 無法登出)\n - Error_ERROR_US_SIGN_ON_FORBIDDEN (上一位使用者沒登出, 無法登入)",
        "operationId": "Operator",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiOperatorReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiOperatorRequest"
            }
          }
        ],
        "tags": [
          "API"
        ]
      }
    },
...
  },
  "definitions": {
...
    "ErrorCode": {
      "type": "string",
      "enum": [
        "ERROR_NONE",
        "ERROR_INTERNAL",
        "ERROR_BAD_REQUEST",
        "ERROR_USER_NOT_FOUND_OR_BAD_PASSWORD",
        "ERROR_USER_NO_PERMISSION",
        "ERROR_USER_MISSING_TOKEN",
        "ERROR_USER_UNKNOWN_TOKEN",
        "ERROR_STATION_NOT_FOUND",
        "ERROR_STATION_ON_HOLD",
        "ERROR_RESOURCE_NOT_FOUND",
        "ERROR_RESOURCE_EXPIRED",
        "ERROR_RESOURCE_ON_HOLD",
        "ERROR_RESOURCE_EXISTED",
        "ERROR_RESOURCE_CONTROL_ABOVE_EXTENDED_COUNT",
        "ERROR_WORKORDER_NOT_FOUND",
        "ERROR_WORKORDER_BAD_BATCH",
        "ERROR_WORKORDER_BAD_STATUS",
        "ERROR_WORKORDER_UNEXPECTED_RESULTS",
        "ERROR_CARRIER_NOT_FOUND",
        "ERROR_CARRIER_IN_USE",
        "ERROR_CARRIER_FULL",
        "ERROR_BATCH_NOT_FOUND",
        "ERROR_BATCH_NOT_READY",
        "ERROR_RECORD_NOT_FOUND",
        "ERROR_RECORD_EXISTED",
        "ERROR_RECIPE_NOT_FOUND",
        "ERROR_US_MISMATCH",
        "ERROR_US_SIGN_ON_FORBIDDEN",
        "ERROR_SR_RESOURCE_IN_USE",
        "ERROR_SR_SITE_NOT_FOUND",
        "ERROR_SW_MISMATCH",
        "ERROR_RW_RESOURCE_NOT_READY",
        "ERROR_RW_QUANTITY_BELOW_MIN",
        "ERROR_RW_QUANTITY_ABOVE_MAX",
        "ERROR_RW_QUANTITY_LOW_DOSAGE",
        "ERROR_RW_QUANTITY_SHORTAGE",
        "ERROR_RW_BAD_GRADE",
        "ERROR_RW_RESOURCE_UNEXPECTED",
        "ERROR_RW_RESOURCE_MISSING"
      ],
      "default": "ERROR_NONE",
      "title": "- ERROR_INTERNAL: code value \u003c 10000 reserved for plant side application, you should use gRPC error instead\nDO NOT USE IT AS MES API ERROR.\n - ERROR_STATION_ON_HOLD: e.g., maintenance.\n - ERROR_RESOURCE_EXISTED: resource id has been used and thus not valid for new resource from Collect\n - ERROR_WORKORDER_BAD_BATCH: the batch is not allowed to be operated. e.g. CLOSED batch can not be closed.\n - ERROR_WORKORDER_BAD_STATUS: work order unexpected status, e.g: load feed work order only allowed ACTIVE status, load collect work order only allowed ACTIVE or CLOSING status.\n - ERROR_WORKORDER_UNEXPECTED_RESULTS: work order unexpected results, e.g: requested todo work order for feeding but return zero-length todo reply.\n - ERROR_CARRIER_IN_USE: carrier in use and cannot be shared\n - ERROR_CARRIER_FULL: carrier can be shared but it's full\n - ERROR_BATCH_NOT_READY: not enough fed batch to collect\n - ERROR_RECORD_NOT_FOUND: not found query record.\n - ERROR_RECORD_EXISTED: record has been collected.\n - ERROR_US_MISMATCH: the user is not the station/site operator.\n - ERROR_US_SIGN_ON_FORBIDDEN: e.g., not qualified/registered to operate the station.\n - ERROR_SR_RESOURCE_IN_USE: the resource is bound to another station/site provided in details.\n - ERROR_SW_MISMATCH: the work order is not being executed the station\n - ERROR_RW_RESOURCE_NOT_READY: details: \"required_preprocess_name\", e.g., \"cutting\"\n - ERROR_RW_QUANTITY_BELOW_MIN: the used quantity is less than the minimum quantity in the recipe.\n - ERROR_RW_QUANTITY_ABOVE_MAX: the used quantity is greater than the maximum quantity in the recipe.\n - ERROR_RW_QUANTITY_LOW_DOSAGE: the used quantity is less than the dosage.\n - ERROR_RW_QUANTITY_SHORTAGE: the quantity of the stock is less than the specified quantity(e.g. used quantity, the minimum quantity in the recipe).\n - ERROR_RW_BAD_GRADE: details: \"required_resource_grade\", e.g., \"AA\"\n - ERROR_RW_RESOURCE_UNEXPECTED: the resource is not in the work order recipe\n - ERROR_RW_RESOURCE_MISSING: the resource is required but missing"
    },
...
    "apiOperatorReply": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/checkError"
        }
      }
    },
    "apiOperatorRequest": {
      "type": "object",
      "properties": {
        "in": {
          "type": "boolean",
          "format": "boolean"
        },
        "group": {
          "type": "integer",
          "format": "int32"
        },
        "workDate": {
          "$ref": "#/definitions/commonsTime"
        },
        "shift": {
          "type": "integer",
          "format": "int32"
        },
        "onlyValidation": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
...
    "checkError": {
      "type": "object",
      "properties": {
        "code": {
          "$ref": "#/definitions/ErrorCode"
        },
        "details": {
          "type": "string"
        }
      }
    },
...
    "commonsTime": {
      "type": "object",
      "properties": {
        "nano": {
          "type": "string",
          "format": "int64",
          "title": "now.UnixNano()"
        }
      }
    },
...
  }
}